convert calendar to its own component
