+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Navigation from '@/components/navigation';
+import Navigation from '@/components/Navigation';
 
 import Image from 'next/image';
